Add basic human-readable Output function
Right now there is no way to call it, that will come once
argparse is added. It will need to include an option to change
the output method as well as suppress SUCCESS.

Very little effort is made to ensure that all possible outputs
are done. The keys for output need to be set into stone still.