Remove implicit_files test, only look at ldapi in net conf output
implicit_files should always be listed but in any case we only
care about the other domains so drop it if present.

There are other options that can change in the Samba configuration
so rather than comparing current vs template config look at
specific options, starting with the ldapi path.