Make input-file globally available
There is value in being able to re-process existing log files for
any output type so move input-file out of Human and process it
in main.

The main difference in processing is where the results come from:
either a file or by running the checks. Post-processing works
the same.

This will allow one to use an existing log to:
- look for only failures
- look at a specific source or check
- output into any format

https://github.com/freeipa/freeipa-healthcheck/issues/50