Add some abstraction to opts parser and output handling
This is primarly for cluster integration. It doesn't need the
json or human output types and providing them will just confuse
users.

It also doesn't need the severity, failures-only and input-file
options, so let those be optional.

This should be backwards compatible with ipa-healthcheck and
any other user of RunChecks since this should default to the
original behavior.