Use new abstracted opts and add Ansible output plugin
This adds a new optional argument to registry: options

The options may be necessary when initializing a registry of
plugins.

Checks will need to return "name" and "value" in the kw results
and these are translated into those names when output.

For example:

yield Result(self, constants.SUCCESS, name='foo', value='bar')

will result in the output:

foo=bar