Filter out the pki healthcheck sources if IPA CA is not installed
The pki checks spew the error "Invalid PKI instance: pki-tomcat" so
we need to suppress them in the IPA CA-less installation case.

So if the IPA CA is not configured then don't register the
pki sources.

A side-effect is that to user the sources will not be listed at
all in this case.

This should not affect pki-healthcheck and it will continue to
return errors in the unconfigured case.

https://github.com/freeipa/freeipa-healthcheck/issues/201

Signed-off-by: Rob Crittenden <rcritten@redhat.com>