Replace or drop print() statements
The print in dogtag/ca.py was supposed to be a debug statement.
Converted it to use logger and fixed a related issue that was
calling logging.debug instead of logger.debug.

ipa/plugin.py didn't assign logger at all and was still using
a direct logging.debug statement.

Drop a print() statement from a test, probably leftover from
troubleshooting a failure.

Signed-off-by: Rob Crittenden <rcritten@redhat.com>

https://github.com/freeipa/freeipa-healthcheck/issues/207