Add keys to some Results that were missing
Being able to suppress a specific key rather than a whole
source or check is better.

I'm not ready yet to assert that there is a key in each
Result since that would be a rather impactful change but
for the purposes of this change I added an assert and ran
it through the unit tests.

https://github.com/freeipa/freeipa-healthcheck/issues/176

Signed-off-by: Rob Crittenden <rcritten@redhat.com>