Read the IPA CA certificate to obtain the serial number
The dogtag connectivity test contains the command
ipa cert-show to verify that the CA basically works
and we are allowed to use it. It had a hardcoded 1 as
the CA certificate since pre-random serial numbers it was
a predictable value.

Instead read the CA cert and pluck the serial number from
it and use that value instead. /etc/ipa/ca.crt may contain
multiple certificates in the case of an external CA install
or additional certs added by ipa-cacert-manage so sift
through the list to find the expected IPA subject.

https://github.com/freeipa/freeipa-healthcheck/issues/260

Signed-off-by: Rob Crittenden <rcritten@redhat.com>